% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettysimplemd.R
\name{prettysimplemd}
\alias{prettysimplemd}
\title{prettysimplemd}
\usage{
prettysimplemd(file, description = NULL, date = NULL, open = NULL, ...)
}
\arguments{
\item{file}{Name of markdown file.}

\item{description}{Optional, short hand description of document contents.}

\item{date}{Date to be timestamped on top left of document. Defaults to 
current year-month-day.}

\item{open}{Logical indicating whether to open output in browser. Defaults to
open the first time the function is called in a session.}

\item{...}{Passed to rmarkdown::render.}
}
\value{
Converts markdown file as pretty simple html file of same root name.
}
\description{
Generates pretty simple markdown files
}
