% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettysimplemd.R
\name{renderPSM}
\alias{renderPSM}
\title{renderPSM}
\usage{
renderPSM(input, open = NULL, ...)
}
\arguments{
\item{input}{Name of file}

\item{open}{Logical indicating whether to open html output in browser}

\item{...}{Passed to \code{\link{prettysimplemd}}}
}
\value{
Output saved as html file.
}
\description{
Renders simple_markdown
}
